Fixed a bug in the download file handler.
As per the VT Private API documentation
(http://www.virustotal.com/en/documentation/private-api/#file-download), the
file/download API does not return a JSON object. This caused the get_response
method to silently throw an exception when it tried to call
jdata = response.json(). This meant that get_response returned response.json
(i.e. a method instance). When jdata's response code was checked back in the
download method, an error would occur because jdata was not a JSON object as
expected, but rather the instance method returned by get_response.